submissions/team-members/sourin-rakshit/requirements.txt
numpy>=1.21
tqdm>=4.64
Pillow>=9.2
kagglehub>=0.3.12
torch>=1.13
torchvision>=0.14
scikit-learn>=1.1
seaborn>=0.11
matplotlib>=3.5